replace internal sound chip clocks with double, same as cpus, and do proper rounding for sample rates, etc
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@5135 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/driver.h
@@ -361,7 +361,7 @@ void machine_remove_sound(struct InternalMachineDriver *machine, const char *tag
 struct InternalMachineDriver
 {
         struct MachineCPU cpu[MAX_CPU];
-        float frames_per_second;
+        double frames_per_second;
         int vblank_duration;
         UINT32 cpu_slices_per_frame;
